Split functionality into separate files
