+++ tui/src/chameleon/commands/hf/__init__.py
@@ -1,0 +1,4 @@
+from chameleon.commands.util import CLITree
+
+
+hf = CLITree("hf", "High Frequency commands", ldr=__loader__)
