+++ tui/src/chameleon/commands/lf/__init__.py
@@ -1,0 +1,4 @@
+from chameleon.commands.util import CLITree
+
+
+lf = CLITree("lf", "Low Frequency commands", ldr=__loader__)
