+++ tui/src/chameleon/commands/lf/em/__init__.py
@@ -1,0 +1,4 @@
+from chameleon.commands.util import CLITree
+
+
+em = CLITree("em", "EM commands", ldr=__loader__)
